+++ tests/uvflag/test_uvflag.py
@@ -1264,7 +1264,7 @@ def test_bad_mode_savefile(uvdata_obj, test_outfile):
 
     # create the file so the clobber gets tested
     with h5py.File(test_outfile, "w") as h5file:
-        h5file.create_dataset("Test", list(range(10)))
+        h5file.create_dataset("Test", list(range(10)), dtype="f4")
 
     uvf.write(test_outfile, clobber=True)
     # manually re-read and tamper with parameters
